requirements.txt
httpx
beautifulsoup4 
lxml
markdownify
pydantic
pydantic-ai
python-dotenv
llama-index
llama-index-embeddings-openai
tqdm
pandas
tiktoken
